Enable hf_transfer for CI test and prefetch
