+++ mlflow/ml-package-versions.yml
@@ -500,6 +500,7 @@ transformers:
       ">= 0.0.0": [
           "datasets",
           "huggingface_hub",
+          "hf_transfer",
           "torch",
           "torchvision",
           "tensorflow",
@@ -526,6 +527,7 @@ transformers:
         [
           "datasets",
           "huggingface_hub",
+          "hf_transfer",
           "torch",
           "torchvision",
           "tensorflow",
