+++ pytest.ini
@@ -4,4 +4,6 @@ filterwarnings =
   # Prevent deprecated numpy type aliases from being used
   error:^`np\.[a-z]+` is a deprecated alias for.+:DeprecationWarning:mlflow
   error:^`np\.[a-z]+` is a deprecated alias for.+:DeprecationWarning:tests
+markers =
+  skipcacheclean: skip cleaning the HuggingFace cache directory after test run, only used for Transformers flavor tests
 timeout = 1200
